+++ CHANGELOG.md
@@ -1,5 +1,22 @@
 # Changelog
 
+## [v0.0.2] - 2026-05-08
+
+### Added
+
+None
+
+### Changed
+
+* Orthomyxoviridae segment numbers automatically mapped to segment names.
+* Dropdown entries are sorted alphabetically
+* Script now prints more status info while running
+
+### Fixed
+
+Tolerate empty figlists when grouping data.
+
+
 ## [v0.0.1] - 2026-05-08
 
 ### Added